refactor: chat 내부 폴더 위치 변경
