+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/(home)/_components/Chat/index.ts
