+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/(home)/_components/ChatDetailLayout/ChatDetailLayout.tsx
