+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/(home)/_components/ChatNotice/index.ts
