+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/(home)/_components/FileButton/index.ts
