+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/(home)/_components/MessageTextarea/index.ts
