+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/(home)/_components/Sidebar/Sidebar.tsx
