+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/(home)/layout.tsx
@@ -1,5 +1,5 @@
 import { StrictPropsWithChildren } from '@/types/common';
-import { ContextMenuProvider } from '../_provider/ContextMenuProvider';
+import { ContextMenuProvider } from './_provider/ContextMenuProvider';
 
 const ChatDetailHomeLayout = ({ children }: StrictPropsWithChildren) => {
   return <ContextMenuProvider>{children}</ContextMenuProvider>;
